Add batting average bot to the repo
A "Twitter Batting Average" is the number of tweets with at least one
retweet or favorite divided by the number of tweets. The tweets
analyzed are the 3200 or fewer most recent tweets of the given user.

This analysis is hooked up to a bot. If a Twitter user tweets
"What is my Twitter batting average?" to the bot, the bot will
compute the average and tweet it back to the user.

This is the first commit of the bot, so it is not very stable.